requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
SQLAlchemy==1.3.23
Flask-SQLAlchemy==2.4.4
Flask-Marshmallow==0.14.0
marshmallow-sqlalchemy==0.23.0
Flask-CORS==3.0.10
python-dotenv==0.19.0
authlib==0.15.3
six==1.16.0
firebase-admin==5.2.0
Jinja2==2.11.3
MarkupSafe==2.0.1
Authlib==2.22